+++ app/models/spina/attachment.rb
@@ -1,19 +1,15 @@
 module Spina
   class Attachment < ApplicationRecord
-    has_one_attached :file
-
+    include Attachable
+    
     attr_accessor :_destroy
 
-    scope :sorted, -> { order('created_at DESC') }
-
-    def name
-      file.filename.to_s
-    end
+    scope :sorted, -> { order("created_at DESC") }
 
     def content
       file if file.attached?
     end
-    
+
     def present?
       signed_blob_id.present?
     end
@@ -21,11 +17,10 @@ module Spina
     alias_method :old_update, :update
     def update(attributes)
       if attributes["_destroy"] == "1" && attributes["file"].blank?
-        self.page_part.destroy
+        page_part.destroy
       else
         old_update(attributes)
       end
     end
-
   end
 end