+++ app/models/spina/image.rb
@@ -1,18 +1,14 @@
 module Spina
   class Image < ApplicationRecord
+    include Attachable
+    
     belongs_to :media_folder, optional: true
 
-    has_one_attached :file
-
-    scope :sorted, -> { order('created_at DESC') }
-
-    def name
-      file.try(:filename).to_s
-    end
+    scope :sorted, -> { order("created_at DESC") }
 
     def variant(options)
       return "" unless file.attached?
-      return file if file.content_type.include?('svg')
+      return file if file.content_type.include?("svg")
       return file unless file.variable?
 
       file.variant(options)
@@ -22,10 +18,10 @@ module Spina
       self
     end
 
-    def thumbnail(size = '100x100', modifier = '^')
+    def thumbnail(size = "100x100", modifier = "^")
       variant(
         combine_options: {
-          gravity: 'center',
+          gravity: "center",
           thumbnail: "#{size}#{modifier}",
           extent: size
         }