+++ app/models/spina/parts/base.rb
@@ -6,14 +6,15 @@ module Spina
       attr_json_config(unknown_key: :strip)
 
       attr_json :name, :string
-      
-      attr_accessor :title, :hint
+
+      attr_accessor :title, :hint, :item_name
 
       def label
         content&.to_s
       end
 
-      def content; end
+      def content
+      end
     end
   end
 end