+++ lib/spina/engine.rb
@@ -1,33 +1,33 @@
-require 'importmap-rails'
-require 'turbo-rails'
-require 'stimulus-rails'
-require 'ancestry'
-require 'breadcrumbs_on_rails'
-require 'kaminari'
-require 'mobility'
-require 'rack-rewrite'
-require 'babosa'
-require 'attr_json'
-require 'view_component'
-require 'jsonapi/serializer'
-require 'browser'
-require 'sprockets/railtie'
+require "importmap-rails"
+require "turbo-rails"
+require "stimulus-rails"
+require "ancestry"
+require "breadcrumbs_on_rails"
+require "kaminari"
+require "mobility"
+require "rack-rewrite"
+require "babosa"
+require "attr_json"
+require "view_component"
+require "jsonapi/serializer"
+require "browser"
+require "sprockets/railtie"
 
 module Spina
   class Engine < ::Rails::Engine
     isolate_namespace Spina
-    
-    config.autoload_paths += %W( #{config.root}/lib )
+
+    config.autoload_paths += %W[#{config.root}/lib]
 
     config.to_prepare do
-      # Require decorators from main application
       unless Spina.config.disable_decorator_load
         Dir.glob(Rails.root + "app/decorators/**/*_decorator.rb").each do |decorator|
+          ActiveSupport::Deprecation.warn("using app/decorators is deprecated in favor of app/overrides. Read more about overriding Spina at spinacms.com/guides")
           require_dependency(decorator)
         end
       end
     end
-    
+
     config.to_prepare do
       Spina::Part.register(
         Spina::Parts::Line,
@@ -37,9 +37,9 @@ module Spina
         Spina::Parts::ImageCollection,
         Spina::Parts::Repeater,
         Spina::Parts::Option,
-        Spina::Parts::Attachment
+        Spina::Parts::Attachment,
+        Spina::Parts::PageLink
       )
     end
-
   end
 end