+++ test/dummy/config/initializers/themes/demo.rb
@@ -1,30 +1,30 @@
 # Theme configuration
 Spina::Theme.register do |theme|
-
-  theme.name = 'demo'
-  theme.title = 'Demo theme'
+  theme.name = "demo"
+  theme.title = "Demo theme"
   theme.resources = [
-    {name: "landing_pages", label: "Landing pages"}, 
-    {name: "blog", label: "Blog"}, 
+    {name: "landing_pages", label: "Landing pages"},
+    {name: "blog", label: "Blog"}
   ]
 
   # All available parts
   theme.parts = [{
-    name: 'repeater',
+    name: "repeater",
     title: "Repeater",
     part_type: "Spina::Parts::Repeater",
-    parts: ['line', 'body', 'image', 'image_collection']
+    parts: ["line", "body", "image", "image_collection"]
   }, {
-    name: 'repeater2',
+    name: "repeater2",
     title: "Repeater",
     part_type: "Spina::Parts::Repeater",
-    parts: ['line', 'image']
+    item_name: "item",
+    parts: ["line", "image"]
   }, {
-    name: 'line',
+    name: "line",
     title: "Line",
     part_type: "Spina::Parts::Line"
   }, {
-    name: 'body',
+    name: "body",
     title: "Body",
     part_type: "Spina::Parts::Text"
   }, {
@@ -32,78 +32,91 @@ Spina::Theme.register do |theme|
     title: "Image collection",
     part_type: "Spina::Parts::ImageCollection"
   }, {
-    name: 'image',
+    name: "image",
     title: "Image",
     part_type: "Spina::Parts::Image"
-  }, {name: 'portrait', part_type: 'Spina::Parts::Image', options: {ratio: "portrait"}
-  }, {name: 'landscape', part_type: 'Spina::Parts::Image', options: {ratio: "landscape"}
-  }, {name: 'wide', part_type: 'Spina::Parts::Image', options: {ratio: "wide"}
-  }, {
-    name: 'headline',
+  }, {name: "portrait", part_type: "Spina::Parts::Image", options: {ratio: "portrait"}}, {name: "landscape", part_type: "Spina::Parts::Image", options: {ratio: "landscape"}}, {name: "wide", part_type: "Spina::Parts::Image", options: {ratio: "wide"}}, {
+    name: "headline",
     title: "Headline",
     hint: "This will be shown in your header",
     part_type: "Spina::Parts::Line"
   }, {
-    name: 'footer',
+    name: "footer",
     title: "Footer",
     part_type: "Spina::Parts::Text"
   }, {
-    name: 'option',
+    name: "option",
     title: "Option",
     part_type: "Spina::Parts::Option",
-    options: [["Left", 'left'], ["Center", 'center'], ["Right", 'right']]
+    options: [["Left", "left"], ["Center", "center"], ["Right", "right"]]
   }, {
-    name: 'attachment',
+    name: "attachment",
     title: "Attachment",
     part_type: "Spina::Parts::Attachment"
   }, {
-    name: 'testrepeater',
-    title: 'Testrepeater',
+    name: "testrepeater",
+    title: "Testrepeater",
     part_type: "Spina::Parts::Repeater",
-    parts: %w(line body)
+    parts: %w[line body]
+  }, {
+    name: "page",
+    title: "Pagina",
+    part_type: "Spina::Parts::PageLink"
+  }, {
+    name: "blogpost",
+    title: "Blogpost",
+    part_type: "Spina::Parts::PageLink",
+    options: {
+      resource: "blog"
+    }
   }]
 
   theme.view_templates = [{
-    name: 'homepage',
-    title: 'Homepage',
-    parts: ['headline', 'body', 'image_collection']
+    name: "homepage",
+    title: "Homepage",
+    parts: ["headline", "body", "image_collection"]
   }, {
-    name: 'show',
-    title: 'Simple page',
+    name: "show",
+    title: "Simple page",
     description: "Default layout",
-    usage: 'Use for your content',
-    parts: ['body', 'testrepeater']
+    usage: "Use for your content",
+    parts: ["body", "blogpost", "page", "testrepeater"]
   }, {
-    name: 'demo',
-    title: 'Demo',
-    description: 'Example including all parts',
-    parts: ['repeater', 'repeater2', 'attachment', 'option', 'body', 'image_collection', 'image', 'portrait', 'landscape', 'wide'],
-    exclude_from: %w(guides)
+    name: "demo",
+    title: "Demo",
+    description: "Example including all parts",
+    parts: ["repeater", "repeater2", "attachment", "option", "body", "image_collection", "image", "portrait", "landscape", "wide"],
+    exclude_from: %w[guides]
   }, {
-    name: 'blogpost',
+    name: "blogpost",
     title: "Blogpost",
-    description: 'Article template',
-    parts: ['body'],
-    exclude_from: %w(main)
+    description: "Article template",
+    parts: ["body"],
+    exclude_from: %w[main]
+  }, {
+    name: "different_layout",
+    title: "Different layout",
+    parts: %w[body],
+    layout: "different_layout"
   }]
 
   theme.custom_pages = [{
-    name:           'homepage',
-    title:          'Homepage',
-    deletable:      false,
-    view_template:  'homepage'
+    name: "homepage",
+    title: "Homepage",
+    deletable: false,
+    view_template: "homepage"
   }, {
-    name:           'demo',
-    title:          'Demo',
-    deletable:      true,
-    view_template:  'demo'
+    name: "demo",
+    title: "Demo",
+    deletable: true,
+    view_template: "demo"
   }]
 
   # Some global content
-  theme.layout_parts = ['line', 'body', 'repeater']
+  theme.layout_parts = ["line", "body", "repeater"]
 
-  theme.plugins = ['reviews']
+  theme.plugins = ["reviews"]
 
   # Embeds
-  theme.embeds = %w(button youtube vimeo)
+  theme.embeds = %w[button youtube vimeo]
 end